After first revue

+++ posts/views.py
@@ -12,12 +12,12 @@ User = get_user_model()
 def index(request):
     post_list = Post.objects.all()
     paginator = Paginator(post_list, 10)
-    page_number = request.GET.get('page')
+    page_number = request.GET.get("page")
     page = paginator.get_page(page_number)
     return render(
         request,
-        'index.html',
-        {'page': page, }
+        "index.html",
+        {"page": page, }
     )
 
 
@@ -26,12 +26,12 @@ def group_posts(request, slug):
 
     group_post_list = group.posts.all()
     paginator = Paginator(group_post_list, 10)
-    page_number = request.GET.get('page')
+    page_number = request.GET.get("page")
     page = paginator.get_page(page_number)
     return render(
         request,
-        'group.html',
-        {'group': group, 'page': page, }
+        "group.html",
+        {"group": group, "page": page, }
     )
 
 
@@ -42,8 +42,8 @@ def new_post(request):
         post = form.save(commit=False)
         post.author = request.user
         form.save()
-        return redirect('index')
-    return render(request, 'new.html', {'form': form})
+        return redirect("index")
+    return render(request, "new.html", {"form": form})
 
 
 def profile(request, username):
@@ -54,20 +54,20 @@ def profile(request, username):
     followers = author.following.all().count()
     following = author.follower.all().count()
     paginator = Paginator(post_list, 10)
-    page_number = request.GET.get('page')
+    page_number = request.GET.get("page")
     page = paginator.get_page(page_number)
     following_status = False
     if user.is_authenticated:
         if Follow.objects.filter(user=user, author=author).exists():
             following_status = True
     return render(
-        request, 'profile.html',
-        {'author': author,
-         'page': page,
-         'posts_amount': posts_amount,
-         'followers': followers,
-         'following': following,
-         'following_status': following_status, }
+        request, "profile.html",
+        {"author": author,
+         "page": page,
+         "posts_amount": posts_amount,
+         "followers": followers,
+         "following": following,
+         "following_status": following_status, }
     )
 
 
@@ -77,21 +77,21 @@ def post_view(request, username, post_id):
     post_list = author.posts.all()
     comments = post.comments.all()
     posts_amount = post_list.count()
-    if request.method == 'POST' and request.user.is_authenticated:
+    if request.method == "POST" and request.user.is_authenticated:
         add_comment(request, username, post_id)
     form = CommentForm()
-    return render(request, 'post.html', {'form': form,
-                                         'post': post,
-                                         'author': author,
-                                         'posts_amount': posts_amount,
-                                         'comments': comments, })
+    return render(request, "post.html", {"form": form,
+                                         "post": post,
+                                         "author": author,
+                                         "posts_amount": posts_amount,
+                                         "comments": comments, })
 
 
 @login_required
 def post_edit(request, username, post_id):
     post = get_object_or_404(Post, pk=post_id, author__username=username)
     if request.user != post.author:
-        return redirect('post', username=username, post_id=post_id)
+        return redirect("post", username=username, post_id=post_id)
     form = PostForm(request.POST or None, files=request.FILES or None,
                     instance=post)
     if form.is_valid():
@@ -99,16 +99,15 @@ def post_edit(request, username, post_id):
         return redirect("post", username=request.user.username,
                         post_id=post_id)
     return render(
-        request, 'new.html', {'form': form, 'post': post,
-                              'is_edited': True},
+        request, "new.html", {"form": form, "post": post,
+                              "is_edited": True},
     )
 
 
 @login_required
 def add_comment(request, username, post_id):
-    author = get_object_or_404(User, username=username)
-    post = get_object_or_404(Post, pk=post_id)
-    post_list = author.posts.all()
+    post = get_object_or_404(Post, pk=post_id, author__username=username)
+    post_list = post.author.posts.all()
     comments = post.comments.all()
     posts_amount = post_list.count()
     form = CommentForm(request.POST)
@@ -117,13 +116,13 @@ def add_comment(request, username, post_id):
         comment.author = request.user
         comment.post = post
         form.save()
-        return redirect("post", username=author.username,
+        return redirect("post", username=post.author.username,
                         post_id=post_id)
-    return render(request, 'post.html', {'form': form,
-                                         'post': post,
-                                         'author': author,
-                                         'posts_amount': posts_amount,
-                                         'comments': comments, })
+    return render(request, 'post.html', {"form": form,
+                                         "post": post,
+                                         "author": post.author,
+                                         "posts_amount": posts_amount,
+                                         "comments": comments, })
 
 
 @login_required
@@ -132,10 +131,10 @@ def follow_index(request):
         author__following__user=request.user)
 
     paginator = Paginator(follow_post_list, 10)
-    page_number = request.GET.get('page')
+    page_number = request.GET.get("page")
     page = paginator.get_page(page_number)
     return render(request, "follow.html", {
-        'page': page, 'paginator': paginator})
+        "page": page, "paginator": paginator})
 
 
 @login_required
@@ -153,7 +152,8 @@ def profile_follow(request, username):
 def profile_unfollow(request, username):
     author = get_object_or_404(User, username=username)
     current_user = request.user
-    Follow.objects.filter(user=current_user, author=author).delete()
+    if Follow.objects.filter(user=current_user, author=author).exists():
+        Follow.objects.filter(user=current_user, author=author).delete()
     return redirect("profile", username=author.username)
